+++ src/react/hooks/__tests__/useMutation.test.tsx
@@ -119,42 +119,85 @@ describe('useMutation Hook', () => {
       expect(result.current.data).toEqual(CREATE_TODO_RESULT);
     });
 
-    it('should ensure the mutation callback function has a stable identity', async () => {
-      const variables = {
-        description: 'Get milk!'
+    it('should ensure the mutation callback function has a stable identity no matter what', async () => {
+      const variables1 = {
+        description: 'Get milk',
+      };
+
+      const data1 = {
+        createTodo: {
+          id: 1,
+          description: 'Get milk!',
+          priority: 'High',
+          __typename: 'Todo',
+        }
+      };
+
+      const variables2 = {
+        description: 'Write blog post',
+      };
+
+      const data2 = {
+        createTodo: {
+          id: 1,
+          description: 'Write blog post',
+          priority: 'High',
+          __typename: 'Todo',
+        },
       };
 
       const mocks = [
         {
           request: {
             query: CREATE_TODO_MUTATION,
-            variables
+            variables: variables1,
           },
-          result: { data: CREATE_TODO_RESULT }
-        }
+          result: { data: data1 },
+        },
+        {
+          request: {
+            query: CREATE_TODO_MUTATION,
+            variables: variables2,
+          },
+          result: { data: data2 },
+        },
       ];
 
-      const { result, waitForNextUpdate } = renderHook(
-        () => useMutation(CREATE_TODO_MUTATION, {}),
-        { wrapper: ({ children }) => (
-          <MockedProvider mocks={mocks}>
-            {children}
-          </MockedProvider>
-        )},
+      const { result, rerender, waitForNextUpdate } = renderHook(
+        ({ variables }) => useMutation(CREATE_TODO_MUTATION, { variables }),
+        {
+          wrapper: ({ children }) => (
+            <MockedProvider mocks={mocks}>
+              {children}
+            </MockedProvider>
+          ),
+          initialProps: {
+            variables: variables1,
+          },
+        },
       );
 
       const createTodo = result.current[0];
       expect(result.current[1].loading).toBe(false);
       expect(result.current[1].data).toBe(undefined);
-      act(() => void createTodo({ variables }));
+
+      act(() => void createTodo());
       expect(createTodo).toBe(result.current[0]);
       expect(result.current[1].loading).toBe(true);
       expect(result.current[1].data).toBe(undefined);
 
       await waitForNextUpdate();
-      expect(createTodo).toBe(result.current[0]);
+      expect(result.current[0]).toBe(createTodo);
       expect(result.current[1].loading).toBe(false);
-      expect(result.current[1].data).toEqual(CREATE_TODO_RESULT);
+      expect(result.current[1].data).toEqual(data1);
+
+      rerender({ variables: variables2 });
+      act(() => void createTodo());
+
+      await waitForNextUpdate();
+      expect(result.current[0]).toBe(createTodo);
+      expect(result.current[1].loading).toBe(false);
+      expect(result.current[1].data).toEqual(data2);
     });
 
     it('should resolve mutate function promise with mutation results', async () => {
@@ -523,6 +566,66 @@ describe('useMutation Hook', () => {
       expect(onCompleted).toHaveBeenCalledTimes(0);
       expect(onError).toHaveBeenCalledTimes(1);
       expect(onError).toHaveBeenCalledWith(errors[0]);
+    });
+
+    it('should never allow onCompleted handler to be stale', async () => {
+      const CREATE_TODO_DATA = {
+        createTodo: {
+          id: 1,
+          priority: 'Low',
+          description: 'Get milk!',
+          __typename: 'Todo',
+        },
+      };
+
+      const mocks = [
+        {
+          request: {
+            query: CREATE_TODO_MUTATION,
+            variables: {
+              priority: 'Low',
+              description: 'Get milk.',
+            }
+          },
+          result: {
+            data: CREATE_TODO_DATA,
+          },
+        }
+      ];
+
+      const onCompleted = jest.fn();
+      const { result, rerender } = renderHook(
+        ({ onCompleted }) => {
+          return useMutation<
+            { createTodo: Todo },
+            { priority: string, description: string }
+          >(CREATE_TODO_MUTATION, { onCompleted });
+        },
+        {
+          wrapper: ({ children }) => (
+            <MockedProvider mocks={mocks}>
+              {children}
+            </MockedProvider>
+          ),
+          initialProps: { onCompleted },
+        },
+      );
+
+      const onCompleted1 = jest.fn();
+      rerender({ onCompleted: onCompleted1 });
+      const createTodo = result.current[0];
+      let fetchResult: any;
+      await act(async () => {
+        fetchResult = await createTodo({
+          variables: { priority: 'Low', description: 'Get milk.' },
+        });
+      });
+
+      expect(fetchResult).toEqual({ data: CREATE_TODO_DATA });
+      expect(result.current[1].data).toEqual(CREATE_TODO_DATA);
+      expect(onCompleted).toHaveBeenCalledTimes(0);
+      expect(onCompleted1).toHaveBeenCalledTimes(1);
+      expect(onCompleted1).toHaveBeenCalledWith(CREATE_TODO_DATA);
     });
   });